+++ src/components/shared/Card/Image/CardImage.vue
@@ -1,14 +1,35 @@
 <template>
-    <img id="card-image" v-lazy="imageUrl" class="card-image small-shadow" alt="Card Image Helper"/>
+    <img id="card-image" :key="imageSource" v-lazy="imageSource" class="card-image small-shadow" alt="Card Image Helper" :style="imageStyle"/>
 </template>
 
 <script>
+import noImage from "@/assets/jpg/no-image.jpg";
+
 export default {
     name: "CardImage",
     props: {
         imageUrl: {
             type: String,
-            required: true,
+            default: "",
+        },
+        maxWidth: {
+            type: Number,
+            default: 100,
+        },
+        maxHeight: {
+            type: Number,
+            default: 100,
+        },
+    },
+    computed: {
+        imageSource() {
+            return this.imageUrl ? this.imageUrl : noImage;
+        },
+        imageStyle() {
+            return {
+                "max-width": `${this.maxWidth}px`,
+                "max-height": `${this.maxHeight}px`,
+            };
         },
     },
 };
@@ -16,8 +37,6 @@ export default {
 
 <style lang="scss" scoped>
     .card-image {
-        max-width: 100px;
-        max-height: 100px;
         border: 1px solid #727272;
         border-radius: 15px;
     }