+++ src/helpers/functions/Date.js
@@ -1,0 +1,7 @@
+export function getFormattedDate(d) {
+    const date = new Date(d);
+    const day = date.getDate().toString().padStart(2, "0");
+    const month = (date.getMonth() + 1).toString().padStart(2, "0");
+    const year = date.getFullYear();
+    return `${day}/${month}/${year}`;
+}