+++ src/plugins/yup/index.js
@@ -1,0 +1,4 @@
+import { setLocale } from "yup";
+import { fr } from "yup-locales";
+
+setLocale(fr);